Bug Fixes: GL Interface: Added [[nodiscard]] to all gl::draw...() functions that returns a vbo pointer. Font Engine: Fixed memory leaks on calling gl::drawTexR8Rectangle(). StringBox: Changed the textureSize to be a uvec2 and that it is ceiling(vec2). ThreadRender: Removed unnessessery const.

+++ Planetery Engine PreBuild v0.3/StringBox.h
@@ -35,7 +35,7 @@ protected:
 	bool _isLineCentre;
 	float _pointSize;
 	vec2 _ppi;
-	vec2 _textureSize;
+	uvec2 _textureSize;
 	gl::Texture2D* _tex;
 	gl::FrameBuffer* _fbo;
 	std::ostringstream _ss;
